+++ api-collector/collector/src/services/__types__/market.types.ts
